+++ doc/changelog.rst
@@ -5,8 +5,8 @@
 Changelog
 =========
 
-2.0 (December 7, 2020)
-----------------------
+2.0.0 (December 7, 2020)
+------------------------
 
 This is a breaking change for iminuit. Interface that was deprecated in 1.x has been removed. In addition, breaking changes were made to the interface to arrive at a clean minimal state that is easier to learn, safer to use, and ready for the long-term future. **To keep existing scripts running, pin your major iminuit version to <2**, i.e. ``pip install 'iminuit<2'`` installs the 1.x series.